requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.4.6
pydantic==2.8.2
python-dotenv==1.0.1
